+++ packages/fxa-auth-server/lib/routes/subscriptions/paypal-notifications.ts
@@ -1,0 +1,38 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+import { ServerRoute } from '@hapi/hapi';
+
+import { ConfigType } from '../../../config';
+import { StripeHelper } from '../../payments/stripe';
+import { AuthLogger } from '../../types';
+import { PayPalHandler } from './paypal';
+
+export class PayPalNotificationHandler extends PayPalHandler {}
+
+export const paypalRoutes = (
+  log: AuthLogger,
+  db: any,
+  config: ConfigType,
+  customs: any,
+  push: any,
+  mailer: any,
+  profile: any,
+  stripeHelper: StripeHelper
+): ServerRoute[] => {
+  const paypalNotificationRoutes = new PayPalNotificationHandler(
+    log,
+    db,
+    config,
+    customs,
+    push,
+    mailer,
+    profile,
+    stripeHelper
+  );
+
+  // FIXME: All of these need to be wrapped in Stripe error handling
+  // FIXME: Many of these stripe calls need retries with careful thought about
+  //        overall request deadline. Stripe retries must include a idempotency_key.
+  return [];
+};
