+++ packages/fxa-auth-server/lib/routes/subscriptions/stripe.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+import { ServerRoute } from '@hapi/hapi';
 import isA from '@hapi/joi';
-import * as Sentry from '@sentry/node';
 import { AbbrevPlan } from 'fxa-shared/dist/subscriptions/types';
 import ScopeSet from 'fxa-shared/oauth/scopes';
 import { metadataFromPlan } from 'fxa-shared/subscriptions/metadata';
@@ -15,28 +15,21 @@ import {
 } from 'fxa-shared/subscriptions/stripe';
 import omitBy from 'lodash/omitBy';
 import { Stripe } from 'stripe';
-import Container from 'typedi';
 
-import { ConfigType } from '../../config';
-import error from '../error';
-import { PayPalHelper } from '../payments/paypal';
-import { StripeHelper, SUBSCRIPTION_UPDATE_TYPES } from '../payments/stripe';
-import { AuthLogger, AuthRequest } from '../types';
-import { splitCapabilities } from './utils/subscriptions';
-import validators from './validators';
+import { ConfigType } from '../../../config';
+import error from '../../error';
+import { StripeHelper } from '../../payments/stripe';
+import { AuthLogger, AuthRequest } from '../../types';
+import { splitCapabilities } from '../utils/subscriptions';
+import validators from '../validators';
 
 const SUBSCRIPTIONS_MANAGEMENT_SCOPE =
   'https://identity.mozilla.com/account/subscriptions';
 
-const IGNORABLE_STRIPE_WEBHOOK_ERRNOS = [
-  error.ERRNO.UNKNOWN_SUBSCRIPTION_FOR_SOURCE,
-  error.ERRNO.BOUNCE_HARD,
-];
-
 /**
  * Authentication handler for subscription routes.
  */
-async function handleAuth(db: any, auth: any, fetchEmail = false) {
+export async function handleAuth(db: any, auth: any, fetchEmail = false) {
   const scope = ScopeSet.fromArray(auth.credentials.scope);
   if (!scope.contains(SUBSCRIPTIONS_MANAGEMENT_SCOPE)) {
     throw error.invalidScopes();
@@ -57,7 +50,7 @@ async function handleAuth(db: any, auth: any, fetchEmail = false) {
  * sending response. We don't need to reveal those.
  * https://github.com/mozilla/fxa/issues/3273#issuecomment-552637420
  */
-function sanitizePlans(plans: AbbrevPlan[]) {
+export function sanitizePlans(plans: AbbrevPlan[]) {
   return plans.map((planIn) => {
     // Try not to mutate the original in case we cache plans in memory.
     const plan = { ...planIn };
@@ -69,21 +62,17 @@ function sanitizePlans(plans: AbbrevPlan[]) {
   });
 }
 
-class DirectStripeRoutes {
-  private paypalHelper: PayPalHelper;
-
+export class StripeHandler {
   constructor(
-    private log: AuthLogger,
-    private db: any,
-    private config: ConfigType,
-    private customs: any,
-    private push: any,
-    private mailer: any,
-    private profile: any,
-    private stripeHelper: StripeHelper
-  ) {
-    this.paypalHelper = Container.get(PayPalHelper);
-  }
+    protected log: AuthLogger,
+    protected db: any,
+    protected config: ConfigType,
+    protected customs: any,
+    protected push: any,
+    protected mailer: any,
+    protected profile: any,
+    protected stripeHelper: StripeHelper
+  ) {}
 
   /**
    * Reload the customer data to reflect a change.
@@ -712,346 +701,6 @@ class DirectStripeRoutes {
   }
 
   /**
-   * Handle webhook events from Stripe by pre-processing the incoming
-   * event and dispatching to the appropriate sub-handler. Log an info
-   * message for events we don't yet handle.
-   */
-  async handleWebhookEvent(request: AuthRequest) {
-    try {
-      const event = this.stripeHelper.constructWebhookEvent(
-        request.payload,
-        request.headers['stripe-signature']
-      );
-
-      switch (event.type) {
-        case 'customer.created':
-          await this.handleCustomerCreatedEvent(request, event);
-          break;
-        case 'customer.subscription.created':
-          await this.handleSubscriptionCreatedEvent(request, event);
-          break;
-        case 'customer.subscription.updated':
-          await this.handleSubscriptionUpdatedEvent(request, event);
-          break;
-        case 'customer.subscription.deleted':
-          await this.handleSubscriptionDeletedEvent(request, event);
-          break;
-        case 'customer.source.expiring':
-          await this.handleCustomerSourceExpiringEvent(request, event);
-          break;
-        case 'invoice.payment_succeeded':
-          await this.handleInvoicePaymentSucceededEvent(request, event);
-          break;
-        case 'invoice.payment_failed':
-          await this.handleInvoicePaymentFailedEvent(request, event);
-          break;
-        default:
-          Sentry.withScope((scope) => {
-            scope.setContext('stripeEvent', {
-              event: { id: event.id, type: event.type },
-            });
-            Sentry.captureMessage(
-              'Unhandled Stripe event received.',
-              Sentry.Severity.Info
-            );
-          });
-          break;
-      }
-    } catch (error) {
-      if (!IGNORABLE_STRIPE_WEBHOOK_ERRNOS.includes(error.errno)) {
-        // Error is not ignorable, so re-throw.
-        throw error;
-      }
-      // Caught an ignorable error, so let's log but continue to 200 OK response
-      this.log.error('subscriptions.handleWebhookEvent.failure', { error });
-    }
-    return {};
-  }
-
-  /**
-   * Handle `customer.created` Stripe webhook events.
-   */
-  async handleCustomerCreatedEvent(_: AuthRequest, event: Stripe.Event) {
-    const customer = event.data.object as Stripe.Customer;
-    const account = await this.db.accountRecord(customer.email);
-    await this.stripeHelper.createLocalCustomer(account.uid, customer);
-  }
-
-  /**
-   * Handle `subscription.created` Stripe webhook events.
-   *
-   * Only subscriptions that are active/trialing are valid. Emit an event for
-   * those conditions only.
-   */
-  async handleSubscriptionCreatedEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const sub = event.data.object as Stripe.Subscription;
-    if (['active', 'trialing'].includes(sub.status)) {
-      return this.updateCustomerAndSendStatus(request, event, sub, true);
-    }
-  }
-
-  /**
-   * Handle `subscription.updated` Stripe webhook events.
-   *
-   * The only time this requires us to emit a subscription event is when an
-   * existing incomplete subscription has now been completed. Unpaid renewals and
-   * subscriptions that are cancelled result in a `subscription.deleted` event.
-   */
-  async handleSubscriptionUpdatedEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const sub = event.data.object as Stripe.Subscription;
-    const { uid, email } = await this.sendSubscriptionUpdatedEmail(event);
-
-    // if the subscription changed from 'incomplete' to 'active' or 'trialing'
-    if (
-      ['active', 'trialing'].includes(sub.status) &&
-      (event.data?.previous_attributes as any)?.status === 'incomplete'
-    ) {
-      return this.updateCustomerAndSendStatus(
-        request,
-        event,
-        sub,
-        true,
-        uid,
-        email
-      );
-    }
-  }
-
-  /**
-   * Handle `subscription.deleted` Stripe wehbook events.
-   */
-  async handleSubscriptionDeletedEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const sub = event.data.object as Stripe.Subscription;
-    const { uid, email } = await this.sendSubscriptionDeletedEmail(sub);
-    return this.updateCustomerAndSendStatus(
-      request,
-      event,
-      sub,
-      false,
-      uid,
-      email
-    );
-  }
-
-  /**
-   * Handle `invoice.payment_succeeded` Stripe wehbook events.
-   */
-  async handleInvoicePaymentSucceededEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const invoice = event.data.object as Stripe.Invoice;
-    const { uid, email } = await this.sendSubscriptionInvoiceEmail(invoice);
-    await this.updateCustomer(uid, email);
-  }
-
-  /**
-   * Handle `invoice.payment_succeeded` Stripe wehbook events.
-   */
-  async handleInvoicePaymentFailedEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const invoice = event.data.object as Stripe.Invoice;
-    if (invoice.billing_reason !== 'subscription_cycle') {
-      // Send payment failure emails only when processing a subscription renewal.
-      return;
-    }
-    const { uid, email } = await this.sendSubscriptionPaymentFailedEmail(
-      invoice
-    );
-    await this.updateCustomer(uid, email);
-  }
-
-  /**
-   * Handle `customer.source.expiring` Stripe wehbook events.
-   */
-  async handleCustomerSourceExpiringEvent(
-    request: AuthRequest,
-    event: Stripe.Event
-  ) {
-    const source = event.data.object as Stripe.Source;
-    const { uid, email } = await this.sendSubscriptionPaymentExpiredEmail(
-      source
-    );
-    await this.updateCustomer(uid, email);
-  }
-
-  /**
-   * Send out an email on payment expiration.
-   */
-  async sendSubscriptionPaymentExpiredEmail(paymentSource: Stripe.Source) {
-    const sourceDetails = await this.stripeHelper.extractSourceDetailsForEmail(
-      paymentSource
-    );
-    const { uid } = sourceDetails;
-    const account = await this.db.account(uid);
-
-    await this.mailer.sendSubscriptionPaymentExpiredEmail(
-      account.emails,
-      account,
-      {
-        acceptLanguage: account.locale,
-        ...sourceDetails,
-      }
-    );
-
-    return sourceDetails;
-  }
-
-  /**
-   * Send out an email on payment failure.
-   */
-  async sendSubscriptionPaymentFailedEmail(invoice: Stripe.Invoice) {
-    const invoiceDetails = await this.stripeHelper.extractInvoiceDetailsForEmail(
-      invoice
-    );
-    const { uid } = invoiceDetails;
-    const account = await this.db.account(uid);
-    await this.mailer.sendSubscriptionPaymentFailedEmail(
-      account.emails,
-      account,
-      {
-        acceptLanguage: account.locale,
-        ...invoiceDetails,
-      }
-    );
-    return invoiceDetails;
-  }
-
-  /**
-   * Send out the appropriate invoice email, depending on whether it's the
-   * initial or a subsequent invoice.
-   */
-  async sendSubscriptionInvoiceEmail(invoice: Stripe.Invoice) {
-    const invoiceDetails = await this.stripeHelper.extractInvoiceDetailsForEmail(
-      invoice
-    );
-    const { uid } = invoiceDetails;
-    const account = await this.db.account(uid);
-    const mailParams = [
-      account.emails,
-      account,
-      {
-        acceptLanguage: account.locale,
-        ...invoiceDetails,
-      },
-    ];
-    switch (invoice.billing_reason) {
-      case 'subscription_create':
-        await this.mailer.sendSubscriptionFirstInvoiceEmail(...mailParams);
-        await this.mailer.sendDownloadSubscriptionEmail(...mailParams);
-        break;
-      default:
-        // Other billing reasons should be covered in subsequent invoice email
-        // https://stripe.com/docs/api/invoices/object#invoice_object-billing_reason
-        await this.mailer.sendSubscriptionSubsequentInvoiceEmail(...mailParams);
-        break;
-    }
-    return invoiceDetails;
-  }
-
-  /**
-   * Send out the appropriate email on subscription update, depending on
-   * whether the change was a subscription upgrade or downgrade.
-   */
-  async sendSubscriptionUpdatedEmail(event: Stripe.Event) {
-    const eventDetails = await this.stripeHelper.extractSubscriptionUpdateEventDetailsForEmail(
-      event
-    );
-    const { uid, email } = eventDetails;
-    const account = await this.db.account(uid);
-
-    const mailParams = [
-      account.emails,
-      account,
-      {
-        acceptLanguage: account.locale,
-        ...eventDetails,
-      },
-    ];
-
-    switch (eventDetails.updateType) {
-      case SUBSCRIPTION_UPDATE_TYPES.UPGRADE:
-        await this.mailer.sendSubscriptionUpgradeEmail(...mailParams);
-        break;
-      case SUBSCRIPTION_UPDATE_TYPES.DOWNGRADE:
-        await this.mailer.sendSubscriptionDowngradeEmail(...mailParams);
-        break;
-      case SUBSCRIPTION_UPDATE_TYPES.REACTIVATION:
-        await this.mailer.sendSubscriptionReactivationEmail(...mailParams);
-        break;
-      case SUBSCRIPTION_UPDATE_TYPES.CANCELLATION:
-        await this.mailer.sendSubscriptionCancellationEmail(...mailParams);
-        break;
-    }
-
-    return { uid, email };
-  }
-
-  /**
-   * Send out the appropriate email on subscription deletion, depending on
-   * whether the user still has an account.
-   */
-  async sendSubscriptionDeletedEmail(subscription: Stripe.Subscription) {
-    if (typeof subscription.latest_invoice !== 'string') {
-      throw error.internalValidationError(
-        'sendSubscriptionDeletedEmail',
-        {
-          subscriptionId: subscription.id,
-          subscriptionInvoiceType: typeof subscription.latest_invoice,
-        },
-        'Subscription latest_invoice was not a string.'
-      );
-    }
-    const invoiceDetails = await this.stripeHelper.extractInvoiceDetailsForEmail(
-      subscription.latest_invoice
-    );
-    if (
-      subscription.metadata &&
-      subscription.metadata.cancelled_for_customer_at
-    ) {
-      // Subscription already cancelled, should have triggered an email earlier
-      return invoiceDetails;
-    }
-    const { uid, email } = invoiceDetails;
-
-    let account;
-    try {
-      // If the user's account has not been deleted, we should have already
-      // sent email at subscription update when cancel_at_period_end = true
-      await this.db.account(uid);
-    } catch (err) {
-      // Has the user's account been deleted?
-      if (err.errno === error.ERRNO.ACCOUNT_UNKNOWN) {
-        // HACK: Minimal account-like object that mailer should accept.
-        // see also: lib/senders/index.js senders.email wrappedMailer
-        account = { email, uid, emails: [{ email, isPrimary: true }] };
-
-        await this.mailer.sendSubscriptionAccountDeletionEmail(
-          account.emails,
-          account,
-          {
-            // TODO: How do we retain account.locale on deletion? Save in customer metadata?
-            // acceptLanguage: account.locale,
-            ...invoiceDetails,
-          }
-        );
-      }
-    }
-    return invoiceDetails;
-  }
-
-  /**
    * Get a list of subscriptions with a FxA UID and email address.
    */
   async getSubscriptions(request: AuthRequest) {
@@ -1111,7 +760,7 @@ class DirectStripeRoutes {
   }
 }
 
-const directRoutes = (
+export const stripeRoutes = (
   log: AuthLogger,
   db: any,
   config: ConfigType,
@@ -1120,8 +769,8 @@ const directRoutes = (
   mailer: any,
   profile: any,
   stripeHelper: StripeHelper
-) => {
-  const directStripeRoutes = new DirectStripeRoutes(
+): ServerRoute[] => {
+  const stripeHandler = new StripeHandler(
     log,
     db,
     config,
@@ -1153,7 +802,7 @@ const directRoutes = (
           ),
         },
       },
-      handler: (request: AuthRequest) => directStripeRoutes.getClients(request),
+      handler: (request: AuthRequest) => stripeHandler.getClients(request),
     },
     {
       method: 'GET',
@@ -1167,7 +816,7 @@ const directRoutes = (
           schema: isA.array().items(validators.subscriptionsPlanValidator),
         },
       },
-      handler: (request: AuthRequest) => directStripeRoutes.listPlans(request),
+      handler: (request: AuthRequest) => stripeHandler.listPlans(request),
     },
     {
       method: 'GET',
@@ -1181,7 +830,7 @@ const directRoutes = (
           schema: isA.array().items(validators.activeSubscriptionValidator),
         },
       },
-      handler: (request: AuthRequest) => directStripeRoutes.listActive(request),
+      handler: (request: AuthRequest) => stripeHandler.listActive(request),
     },
     {
       method: 'GET',
@@ -1195,8 +844,7 @@ const directRoutes = (
           schema: validators.subscriptionsCustomerValidator,
         },
       },
-      handler: (request: AuthRequest) =>
-        directStripeRoutes.getCustomer(request),
+      handler: (request: AuthRequest) => stripeHandler.getCustomer(request),
     },
     {
       method: 'POST',
@@ -1216,8 +864,7 @@ const directRoutes = (
           },
         },
       },
-      handler: (request: AuthRequest) =>
-        directStripeRoutes.createCustomer(request),
+      handler: (request: AuthRequest) => stripeHandler.createCustomer(request),
     },
     {
       method: 'POST',
@@ -1244,7 +891,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.createSubscriptionWithPMI(request),
+        stripeHandler.createSubscriptionWithPMI(request),
     },
     {
       method: 'POST',
@@ -1265,8 +912,7 @@ const directRoutes = (
           },
         },
       },
-      handler: (request: AuthRequest) =>
-        directStripeRoutes.retryInvoice(request),
+      handler: (request: AuthRequest) => stripeHandler.retryInvoice(request),
     },
     {
       method: 'POST',
@@ -1281,7 +927,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.createSetupIntent(request),
+        stripeHandler.createSetupIntent(request),
     },
     {
       method: 'POST',
@@ -1301,7 +947,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.updateDefaultPaymentMethod(request),
+        stripeHandler.updateDefaultPaymentMethod(request),
     },
     {
       method: 'POST',
@@ -1325,7 +971,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.detachFailedPaymentMethod(request),
+        stripeHandler.detachFailedPaymentMethod(request),
     },
     {
       method: 'GET',
@@ -1342,8 +988,7 @@ const directRoutes = (
           },
         },
       },
-      handler: (request: AuthRequest) =>
-        directStripeRoutes.getProductName(request),
+      handler: (request: AuthRequest) => stripeHandler.getProductName(request),
     },
     {
       method: 'GET',
@@ -1367,7 +1012,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.getSubscriptionsForSupport(request),
+        stripeHandler.getSubscriptionsForSupport(request),
     },
     {
       method: 'PUT',
@@ -1387,7 +1032,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.updateSubscription(request),
+        stripeHandler.updateSubscription(request),
     },
     {
       method: 'DELETE',
@@ -1404,7 +1049,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.deleteSubscription(request),
+        stripeHandler.deleteSubscription(request),
     },
     {
       method: 'POST',
@@ -1421,62 +1066,7 @@ const directRoutes = (
         },
       },
       handler: (request: AuthRequest) =>
-        directStripeRoutes.reactivateSubscription(request),
-    },
-    {
-      method: 'POST',
-      path: '/oauth/subscriptions/stripe/event',
-      options: {
-        // We'll use the official Stripe library to authenticate the payload,
-        // and it will also return an event.
-        auth: false,
-        // The raw payload is needed for authentication.
-        payload: {
-          output: 'data',
-          parse: false,
-        },
-        validate: {
-          headers: { 'stripe-signature': isA.string().required() },
-        },
-      },
-      handler: (request: AuthRequest) =>
-        directStripeRoutes.handleWebhookEvent(request),
+        stripeHandler.reactivateSubscription(request),
     },
   ];
 };
-
-const createRoutes = (
-  log: AuthLogger,
-  db: any,
-  config: ConfigType,
-  customs: any,
-  push: any,
-  mailer: any,
-  profile: any,
-  stripeHelper: StripeHelper
-) => {
-  // Skip routes if the subscriptions feature is not configured & enabled
-  if (!config.subscriptions || !config.subscriptions.enabled) {
-    return [];
-  }
-
-  if (stripeHelper) {
-    return directRoutes(
-      log,
-      db,
-      config,
-      customs,
-      push,
-      mailer,
-      profile,
-      stripeHelper
-    );
-  }
-
-  return [];
-};
-
-module.exports = createRoutes;
-module.exports.DirectStripeRoutes = DirectStripeRoutes;
-module.exports.handleAuth = handleAuth;
-module.exports.sanitizePlans = sanitizePlans;
