+++ packages/fxa-auth-server/test/local/routes/subscriptions/stripe.js
@@ -7,23 +7,18 @@
 const sinon = require('sinon');
 const assert = { ...sinon.assert, ...require('chai').assert };
 const uuid = require('uuid');
-const { getRoute } = require('../../routes_helpers');
-const mocks = require('../../mocks');
-const error = require('../../../lib/error');
-const Sentry = require('@sentry/node');
-const {
-  StripeHelper,
-  SUBSCRIPTION_UPDATE_TYPES,
-} = require('../../../lib/payments/stripe');
+const { getRoute } = require('../../../routes_helpers');
+const mocks = require('../../../mocks');
+const error = require('../../../../lib/error');
+const { StripeHelper } = require('../../../../lib/payments/stripe');
 const WError = require('verror').WError;
 const uuidv4 = require('uuid').v4;
-const moment = require('moment');
 
 const {
   sanitizePlans,
   handleAuth,
   DirectStripeRoutes,
-} = require('../../../lib/routes/subscriptions');
+} = require('../../../../lib/routes/subscriptions');
 
 const {
   filterCustomer,
@@ -32,26 +27,18 @@ const {
   filterIntent,
 } = require('fxa-shared').subscriptions.stripe;
 
-const subscription2 = require('../payments/fixtures/subscription2.json');
-const cancelledSubscription = require('../payments/fixtures/subscription_cancelled.json');
-const trialSubscription = require('../payments/fixtures/subscription_trialing.json');
-const pastDueSubscription = require('../payments/fixtures/subscription_past_due.json');
-const customerFixture = require('../payments/fixtures/customer1.json');
-const customerPMIExpanded = require('../payments/fixtures/customer_new_pmi_default_invoice_expanded.json');
-const multiPlanSubscription = require('../payments/fixtures/subscription_multiplan.json');
-const emptyCustomer = require('../payments/fixtures/customer_new.json');
-const subscriptionCreated = require('../payments/fixtures/subscription_created.json');
-const subscriptionCreatedIncomplete = require('../payments/fixtures/subscription_created_incomplete.json');
-const subscriptionDeleted = require('../payments/fixtures/subscription_deleted.json');
-const subscriptionUpdated = require('../payments/fixtures/subscription_updated.json');
-const subscriptionUpdatedFromIncomplete = require('../payments/fixtures/subscription_updated_from_incomplete.json');
-const eventInvoicePaymentSucceeded = require('../payments/fixtures/event_invoice_payment_succeeded.json');
-const eventInvoicePaymentFailed = require('../payments/fixtures/event_invoice_payment_failed.json');
-const eventCustomerSubscriptionUpdated = require('../payments/fixtures/event_customer_subscription_updated.json');
-const eventCustomerSourceExpiring = require('../payments/fixtures/event_customer_source_expiring.json');
-const openInvoice = require('../payments/fixtures/invoice_open.json');
-const newSetupIntent = require('../payments/fixtures/setup_intent_new.json');
-const stripePlan = require('../payments/fixtures/plan1.json');
+const subscription2 = require('../../payments/fixtures/subscription2.json');
+const cancelledSubscription = require('../../payments/fixtures/subscription_cancelled.json');
+const trialSubscription = require('../../payments/fixtures/subscription_trialing.json');
+const pastDueSubscription = require('../../payments/fixtures/subscription_past_due.json');
+const customerFixture = require('../../payments/fixtures/customer1.json');
+const customerPMIExpanded = require('../../payments/fixtures/customer_new_pmi_default_invoice_expanded.json');
+const multiPlanSubscription = require('../../payments/fixtures/subscription_multiplan.json');
+const emptyCustomer = require('../../payments/fixtures/customer_new.json');
+const subscriptionUpdatedFromIncomplete = require('../../payments/fixtures/subscription_updated_from_incomplete.json');
+const openInvoice = require('../../payments/fixtures/invoice_open.json');
+const newSetupIntent = require('../../payments/fixtures/setup_intent_new.json');
+const stripePlan = require('../../payments/fixtures/plan1.json');
 
 let config,
   log,
@@ -132,7 +119,7 @@ const MOCK_TTL = 3600;
 const MOCK_SCOPES = ['profile:email', SUBSCRIPTIONS_MANAGEMENT_SCOPE];
 
 function runTest(routePath, requestOptions, payments = null) {
-  routes = require('../../../lib/routes/subscriptions')(
+  routes = require('../../../../lib/routes/subscriptions')(
     log,
     db,
     config,
@@ -202,9 +189,9 @@ describe('sanitizePlans', () => {
 });
 
 /**
- * Direct Stripe integration tests
+ * Stripe integration tests
  */
-describe('subscriptions directRoutes', () => {
+describe('subscriptions stripeRoutes', () => {
   beforeEach(() => {
     config = {
       subscriptions: {
@@ -1560,696 +1547,6 @@ describe('DirectStripeRoutes', () => {
         assert.notCalled(profile.deleteCache);
         assert.notCalled(
           directStripeRoutesInstance.sendSubscriptionStatusToSqs
-        );
-      });
-    });
-  });
-
-  describe('stripe webhooks', () => {
-    let stubSendSubscriptionStatusToSqs;
-
-    beforeEach(() => {
-      directStripeRoutesInstance.stripeHelper.getCustomerUidEmailFromSubscription.resolves(
-        { uid: UID, email: TEST_EMAIL }
-      );
-      stubSendSubscriptionStatusToSqs = sandbox
-        .stub(directStripeRoutesInstance, 'sendSubscriptionStatusToSqs')
-        .resolves(true);
-    });
-
-    describe('handleWebhookEvent', () => {
-      let scopeContextSpy, scopeSpy;
-      const request = {
-        payload: {},
-        headers: {
-          'stripe-signature': 'stripe_123',
-        },
-      };
-      const handlerNames = [
-        'handleCustomerCreatedEvent',
-        'handleSubscriptionCreatedEvent',
-        'handleSubscriptionUpdatedEvent',
-        'handleSubscriptionDeletedEvent',
-        'handleCustomerSourceExpiringEvent',
-        'handleInvoicePaymentSucceededEvent',
-        'handleInvoicePaymentFailedEvent',
-      ];
-      const handlerStubs = {};
-
-      beforeEach(() => {
-        for (const handlerName of handlerNames) {
-          handlerStubs[handlerName] = sandbox
-            .stub(directStripeRoutesInstance, handlerName)
-            .resolves();
-        }
-        scopeContextSpy = sinon.fake();
-        scopeSpy = {
-          setContext: scopeContextSpy,
-        };
-        sandbox.replace(Sentry, 'withScope', (fn) => fn(scopeSpy));
-      });
-
-      const assertNamedHandlerCalled = (expectedHandlerName = null) => {
-        for (const handlerName of handlerNames) {
-          const shouldCall =
-            expectedHandlerName && handlerName === expectedHandlerName;
-          assert.isTrue(
-            handlerStubs[handlerName][shouldCall ? 'called' : 'notCalled'],
-            `Expected to ${shouldCall ? '' : 'not '}call ${handlerName}`
-          );
-        }
-      };
-
-      const itOnlyCallsThisHandler = (expectedHandlerName, event) =>
-        it(`only calls ${expectedHandlerName}`, async () => {
-          const createdEvent = deepCopy(event);
-          directStripeRoutesInstance.stripeHelper.constructWebhookEvent.returns(
-            createdEvent
-          );
-          await directStripeRoutesInstance.handleWebhookEvent(request);
-          assertNamedHandlerCalled(expectedHandlerName);
-          assert.isTrue(
-            scopeContextSpy.notCalled,
-            'Expected to not call Sentry'
-          );
-        });
-
-      describe('ignorable errors', () => {
-        const commonIgnorableErrorTest = (expectedError) => async () => {
-          const fixture = deepCopy(eventCustomerSourceExpiring);
-          handlerStubs.handleCustomerSourceExpiringEvent.throws(expectedError);
-          directStripeRoutesInstance.stripeHelper.constructWebhookEvent.returns(
-            fixture
-          );
-          let errorThrown = null;
-          try {
-            await directStripeRoutesInstance.handleWebhookEvent(request);
-            assert.calledWith(
-              directStripeRoutesInstance.log.error,
-              'subscriptions.handleWebhookEvent.failure',
-              { error: expectedError }
-            );
-          } catch (err) {
-            errorThrown = err;
-          }
-          assert.isNull(errorThrown);
-        };
-
-        it(
-          'ignores emailBouncedHard',
-          commonIgnorableErrorTest(error.emailBouncedHard(42))
-        );
-
-        it(
-          'ignores missingSubscriptionForSourceError',
-          commonIgnorableErrorTest(
-            error.missingSubscriptionForSourceError(
-              'extractSourceDetailsForEmail'
-            )
-          )
-        );
-      });
-
-      describe('when the event.type is customer.created', () => {
-        itOnlyCallsThisHandler('handleCustomerCreatedEvent', {
-          data: { object: customerFixture },
-          type: 'customer.created',
-        });
-      });
-
-      describe('when the event.type is customer.subscription.created', () => {
-        itOnlyCallsThisHandler(
-          'handleSubscriptionCreatedEvent',
-          subscriptionCreated
-        );
-      });
-
-      describe('when the event.type is customer.subscription.updated', () => {
-        itOnlyCallsThisHandler(
-          'handleSubscriptionUpdatedEvent',
-          subscriptionUpdated
-        );
-      });
-
-      describe('when the event.type is customer.subscription.deleted', () => {
-        itOnlyCallsThisHandler(
-          'handleSubscriptionDeletedEvent',
-          subscriptionDeleted
-        );
-      });
-
-      describe('when the event.type is customer.source.expiring', () => {
-        itOnlyCallsThisHandler(
-          'handleCustomerSourceExpiringEvent',
-          eventCustomerSourceExpiring
-        );
-      });
-
-      describe('when the event.type is invoice.payment_succeeded', () => {
-        itOnlyCallsThisHandler(
-          'handleInvoicePaymentSucceededEvent',
-          eventInvoicePaymentSucceeded
-        );
-      });
-
-      describe('when the event.type is invoice.payment_failed', () => {
-        itOnlyCallsThisHandler(
-          'handleInvoicePaymentFailedEvent',
-          eventInvoicePaymentFailed
-        );
-      });
-
-      describe('when the event.type is something else', () => {
-        it('only calls sentry', async () => {
-          const event = deepCopy(subscriptionCreated);
-          event.type = 'customer.updated';
-          directStripeRoutesInstance.stripeHelper.constructWebhookEvent.returns(
-            event
-          );
-          await directStripeRoutesInstance.handleWebhookEvent(request);
-          assertNamedHandlerCalled();
-          assert.isTrue(scopeContextSpy.calledOnce, 'Expected to call Sentry');
-        });
-      });
-    });
-
-    const assertSendSubscriptionStatusToSqsCalledWith = (event, isActive) =>
-      assert.calledWith(
-        stubSendSubscriptionStatusToSqs,
-        {},
-        UID,
-        event,
-        { id: event.data.object.id, productId: event.data.object.plan.product },
-        isActive
-      );
-
-    describe('handleCustomerCreatedEvent', () => {
-      it('creates a local db record with the account uid', async () => {
-        await directStripeRoutesInstance.handleCustomerCreatedEvent(
-          {},
-          {
-            data: { object: customerFixture },
-            type: 'customer.created',
-          }
-        );
-
-        assert.calledOnceWithExactly(
-          directStripeRoutesInstance.db.accountRecord,
-          customerFixture.email
-        );
-        assert.calledOnceWithExactly(
-          directStripeRoutesInstance.stripeHelper.createLocalCustomer,
-          UID,
-          customerFixture
-        );
-      });
-    });
-
-    describe('handleSubscriptionUpdatedEvent', () => {
-      let sendSubscriptionUpdatedEmailStub;
-
-      beforeEach(() => {
-        sendSubscriptionUpdatedEmailStub = sandbox
-          .stub(directStripeRoutesInstance, 'sendSubscriptionUpdatedEmail')
-          .resolves({ uid: UID, email: TEST_EMAIL });
-      });
-
-      it('emits a notification when transitioning from "incomplete" to "active/trialing"', async () => {
-        const updatedEvent = deepCopy(subscriptionUpdatedFromIncomplete);
-        await directStripeRoutesInstance.handleSubscriptionUpdatedEvent(
-          {},
-          updatedEvent
-        );
-        assert.calledWith(sendSubscriptionUpdatedEmailStub, updatedEvent);
-        assert.calledWith(
-          directStripeRoutesInstance.stripeHelper.refreshCachedCustomer,
-          UID,
-          TEST_EMAIL
-        );
-        assert.calledWith(profile.deleteCache, UID);
-        assertSendSubscriptionStatusToSqsCalledWith(updatedEvent, true);
-      });
-
-      it('does not emit a notification for any other subscription state change', async () => {
-        const updatedEvent = deepCopy(subscriptionUpdated);
-        await directStripeRoutesInstance.handleSubscriptionUpdatedEvent(
-          {},
-          updatedEvent
-        );
-        assert.calledWith(sendSubscriptionUpdatedEmailStub, updatedEvent);
-        assert.notCalled(
-          directStripeRoutesInstance.stripeHelper.refreshCachedCustomer
-        );
-        assert.notCalled(profile.deleteCache);
-        assert.notCalled(stubSendSubscriptionStatusToSqs);
-      });
-    });
-
-    describe('handleSubscriptionDeletedEvent', () => {
-      it('sends email and emits a notification when a subscription is deleted', async () => {
-        const deletedEvent = deepCopy(subscriptionDeleted);
-        const sendSubscriptionDeletedEmailStub = sandbox
-          .stub(directStripeRoutesInstance, 'sendSubscriptionDeletedEmail')
-          .resolves({ uid: UID, email: TEST_EMAIL });
-        await directStripeRoutesInstance.handleSubscriptionDeletedEvent(
-          {},
-          deletedEvent
-        );
-        assert.calledWith(
-          sendSubscriptionDeletedEmailStub,
-          deletedEvent.data.object
-        );
-        assert.notCalled(
-          directStripeRoutesInstance.stripeHelper
-            .getCustomerUidEmailFromSubscription
-        );
-        assert.calledWith(
-          directStripeRoutesInstance.stripeHelper.refreshCachedCustomer,
-          UID,
-          TEST_EMAIL
-        );
-        assert.calledWith(profile.deleteCache, UID);
-        assertSendSubscriptionStatusToSqsCalledWith(deletedEvent, false);
-      });
-    });
-
-    describe('handleInvoicePaymentSucceededEvent', () => {
-      it('sends email and emits a notification when an invoice payment succeeds', async () => {
-        const paymentSucceededEvent = deepCopy(eventInvoicePaymentSucceeded);
-        const sendSubscriptionInvoiceEmailStub = sandbox
-          .stub(directStripeRoutesInstance, 'sendSubscriptionInvoiceEmail')
-          .resolves(true);
-        const mockSubscription = {
-          id: 'test1',
-          plan: { product: 'test2' },
-        };
-        directStripeRoutesInstance.stripeHelper.expandResource.resolves(
-          mockSubscription
-        );
-        await directStripeRoutesInstance.handleInvoicePaymentSucceededEvent(
-          {},
-          paymentSucceededEvent
-        );
-        assert.calledWith(
-          sendSubscriptionInvoiceEmailStub,
-          paymentSucceededEvent.data.object
-        );
-        assert.notCalled(stubSendSubscriptionStatusToSqs);
-      });
-    });
-
-    describe('handleInvoicePaymentFailedEvent', () => {
-      const mockSubscription = {
-        id: 'test1',
-        plan: { product: 'test2' },
-      };
-      let sendSubscriptionPaymentFailedEmailStub;
-
-      beforeEach(() => {
-        sendSubscriptionPaymentFailedEmailStub = sandbox
-          .stub(
-            directStripeRoutesInstance,
-            'sendSubscriptionPaymentFailedEmail'
-          )
-          .resolves(true);
-        directStripeRoutesInstance.stripeHelper.expandResource.resolves(
-          mockSubscription
-        );
-      });
-
-      it('sends email and emits a notification when an invoice payment fails', async () => {
-        const paymentFailedEvent = deepCopy(eventInvoicePaymentFailed);
-        paymentFailedEvent.data.object.billing_reason = 'subscription_cycle';
-        await directStripeRoutesInstance.handleInvoicePaymentFailedEvent(
-          {},
-          paymentFailedEvent
-        );
-        assert.calledWith(
-          sendSubscriptionPaymentFailedEmailStub,
-          paymentFailedEvent.data.object
-        );
-        assert.notCalled(stubSendSubscriptionStatusToSqs);
-      });
-
-      it('does not send email during subscription creation flow', async () => {
-        const paymentFailedEvent = deepCopy(eventInvoicePaymentFailed);
-        paymentFailedEvent.data.object.billing_reason = 'subscription_create';
-        await directStripeRoutesInstance.handleInvoicePaymentFailedEvent(
-          {},
-          paymentFailedEvent
-        );
-        assert.notCalled(sendSubscriptionPaymentFailedEmailStub);
-      });
-    });
-
-    describe('handleSubscriptionCreatedEvent', () => {
-      it('emits a notification when a new subscription is "active" or "trialing"', async () => {
-        const createdEvent = deepCopy(subscriptionCreated);
-        await directStripeRoutesInstance.handleSubscriptionCreatedEvent(
-          {},
-          createdEvent
-        );
-        assert.calledWith(
-          directStripeRoutesInstance.stripeHelper
-            .getCustomerUidEmailFromSubscription,
-          createdEvent.data.object
-        );
-        assert.calledWith(
-          directStripeRoutesInstance.stripeHelper.refreshCachedCustomer,
-          UID,
-          TEST_EMAIL
-        );
-        assert.calledWith(profile.deleteCache, UID);
-        assertSendSubscriptionStatusToSqsCalledWith(createdEvent, true);
-      });
-
-      it('does not emit a notification for incomplete new subscriptions', async () => {
-        const createdEvent = deepCopy(subscriptionCreatedIncomplete);
-        await directStripeRoutesInstance.handleSubscriptionCreatedEvent(
-          {},
-          createdEvent
-        );
-        assert.notCalled(
-          directStripeRoutesInstance.stripeHelper
-            .getCustomerUidEmailFromSubscription
-        );
-        assert.notCalled(
-          directStripeRoutesInstance.stripeHelper.refreshCachedCustomer
-        );
-        assert.notCalled(profile.deleteCache);
-        assert.notCalled(stubSendSubscriptionStatusToSqs);
-      });
-    });
-  });
-
-  describe('sendSubscriptionPaymentExpiredEmail', () => {
-    const mockSource = {};
-    const mockAccount = {
-      emails: TEST_EMAIL,
-      locale: ACCOUNT_LOCALE,
-    };
-
-    it('sends the email with a list of subscriptions', async () => {
-      directStripeRoutesInstance.db.account = sandbox
-        .stub()
-        .resolves(mockAccount);
-      directStripeRoutesInstance.mailer.sendMultiSubscriptionsPaymentExpiredEmail = sandbox.stub();
-      directStripeRoutesInstance.mailer.sendSubscriptionPaymentExpiredEmail = sandbox.stub();
-      const mockCustomer = { uid: UID, subscriptions: [{ id: 'sub_testo' }] };
-      directStripeRoutesInstance.stripeHelper.extractSourceDetailsForEmail.resolves(
-        mockCustomer
-      );
-
-      await directStripeRoutesInstance.sendSubscriptionPaymentExpiredEmail(
-        mockSource
-      );
-
-      assert.calledOnceWithExactly(
-        directStripeRoutesInstance.stripeHelper.extractSourceDetailsForEmail,
-        mockSource
-      );
-      assert.calledOnceWithExactly(directStripeRoutesInstance.db.account, UID);
-      sinon.assert.calledOnceWithExactly(
-        directStripeRoutesInstance.mailer.sendSubscriptionPaymentExpiredEmail,
-        TEST_EMAIL,
-        { emails: TEST_EMAIL, locale: ACCOUNT_LOCALE },
-        {
-          acceptLanguage: ACCOUNT_LOCALE,
-          ...mockCustomer,
-        }
-      );
-    });
-  });
-
-  describe('sendSubscriptionPaymentFailedEmail', () => {
-    it('sends the payment failed email', async () => {
-      const invoice = deepCopy(eventInvoicePaymentFailed.data.object);
-
-      const mockInvoiceDetails = { uid: '1234', test: 'fake' };
-      directStripeRoutesInstance.stripeHelper.extractInvoiceDetailsForEmail.resolves(
-        mockInvoiceDetails
-      );
-
-      const mockAccount = { emails: 'fakeemails', locale: 'fakelocale' };
-      directStripeRoutesInstance.db.account = sinon.spy(
-        async (data) => mockAccount
-      );
-
-      await directStripeRoutesInstance.sendSubscriptionPaymentFailedEmail(
-        invoice
-      );
-      assert.calledWith(
-        directStripeRoutesInstance.mailer.sendSubscriptionPaymentFailedEmail,
-        mockAccount.emails,
-        mockAccount,
-        {
-          acceptLanguage: mockAccount.locale,
-          ...mockInvoiceDetails,
-        }
-      );
-    });
-  });
-
-  describe('sendSubscriptionInvoiceEmail', () => {
-    const commonSendSubscriptionInvoiceEmailTest = (
-      expectedMethodName,
-      billingReason
-    ) => async () => {
-      const invoice = deepCopy(eventInvoicePaymentSucceeded.data.object);
-      invoice.billing_reason = billingReason;
-
-      const mockInvoiceDetails = { uid: '1234', test: 'fake' };
-      directStripeRoutesInstance.stripeHelper.extractInvoiceDetailsForEmail.resolves(
-        mockInvoiceDetails
-      );
-
-      const mockAccount = { emails: 'fakeemails', locale: 'fakelocale' };
-      directStripeRoutesInstance.db.account = sinon.spy(
-        async (data) => mockAccount
-      );
-
-      await directStripeRoutesInstance.sendSubscriptionInvoiceEmail(invoice);
-      assert.calledWith(
-        directStripeRoutesInstance.mailer[expectedMethodName],
-        mockAccount.emails,
-        mockAccount,
-        {
-          acceptLanguage: mockAccount.locale,
-          ...mockInvoiceDetails,
-        }
-      );
-      if (expectedMethodName === 'sendSubscriptionFirstInvoiceEmail') {
-        assert.calledWith(
-          directStripeRoutesInstance.mailer.sendDownloadSubscriptionEmail,
-          mockAccount.emails,
-          mockAccount,
-          {
-            acceptLanguage: mockAccount.locale,
-            ...mockInvoiceDetails,
-          }
-        );
-      }
-    };
-
-    it(
-      'sends the initial invoice email for a newly created subscription',
-      commonSendSubscriptionInvoiceEmailTest(
-        'sendSubscriptionFirstInvoiceEmail',
-        'subscription_create'
-      )
-    );
-
-    it(
-      'sends the subsequent invoice email for billing reasons besides creation',
-      commonSendSubscriptionInvoiceEmailTest(
-        'sendSubscriptionSubsequentInvoiceEmail',
-        'subscription_cycle'
-      )
-    );
-  });
-
-  describe('sendSubscriptionUpdatedEmail', () => {
-    const commonSendSubscriptionUpdatedEmailTest = (updateType) => async () => {
-      const event = deepCopy(eventCustomerSubscriptionUpdated);
-
-      const mockDetails = {
-        uid: '1234',
-        test: 'fake',
-        updateType,
-      };
-      directStripeRoutesInstance.stripeHelper.extractSubscriptionUpdateEventDetailsForEmail.resolves(
-        mockDetails
-      );
-
-      const mockAccount = { emails: 'fakeemails', locale: 'fakelocale' };
-      directStripeRoutesInstance.db.account = sinon.spy(
-        async (data) => mockAccount
-      );
-
-      await directStripeRoutesInstance.sendSubscriptionUpdatedEmail(event);
-
-      const expectedMethodName = {
-        [SUBSCRIPTION_UPDATE_TYPES.UPGRADE]: 'sendSubscriptionUpgradeEmail',
-        [SUBSCRIPTION_UPDATE_TYPES.DOWNGRADE]: 'sendSubscriptionDowngradeEmail',
-        [SUBSCRIPTION_UPDATE_TYPES.REACTIVATION]:
-          'sendSubscriptionReactivationEmail',
-        [SUBSCRIPTION_UPDATE_TYPES.CANCELLATION]:
-          'sendSubscriptionCancellationEmail',
-      }[updateType];
-
-      assert.calledWith(
-        directStripeRoutesInstance.mailer[expectedMethodName],
-        mockAccount.emails,
-        mockAccount,
-        {
-          acceptLanguage: mockAccount.locale,
-          ...mockDetails,
-        }
-      );
-    };
-
-    it(
-      'sends an upgrade email on subscription upgrade',
-      commonSendSubscriptionUpdatedEmailTest(SUBSCRIPTION_UPDATE_TYPES.UPGRADE)
-    );
-
-    it(
-      'sends a downgrade email on subscription downgrade',
-      commonSendSubscriptionUpdatedEmailTest(
-        SUBSCRIPTION_UPDATE_TYPES.DOWNGRADE
-      )
-    );
-
-    it(
-      'sends a reactivation email on subscription reactivation',
-      commonSendSubscriptionUpdatedEmailTest(
-        SUBSCRIPTION_UPDATE_TYPES.REACTIVATION
-      )
-    );
-
-    it(
-      'sends a cancellation email on subscription cancellation',
-      commonSendSubscriptionUpdatedEmailTest(
-        SUBSCRIPTION_UPDATE_TYPES.CANCELLATION
-      )
-    );
-  });
-
-  describe('sendSubscriptionDeletedEmail', () => {
-    const commonSendSubscriptionDeletedEmailTest = (
-      accountFound = true,
-      subscriptionAlreadyCancelled = false
-    ) => async () => {
-      const deletedEvent = deepCopy(subscriptionDeleted);
-      const subscription = deletedEvent.data.object;
-
-      if (subscriptionAlreadyCancelled) {
-        subscription.metadata = {
-          cancelled_for_customer_at: moment().unix(),
-        };
-      }
-
-      const mockInvoiceDetails = {
-        uid: '1234',
-        test: 'fake',
-        email: 'test@example.com',
-      };
-      directStripeRoutesInstance.stripeHelper.extractInvoiceDetailsForEmail.resolves(
-        mockInvoiceDetails
-      );
-
-      const mockAccount = { emails: 'fakeemails', locale: 'fakelocale' };
-      directStripeRoutesInstance.db.account = sinon.spy(async (data) => {
-        if (accountFound) {
-          return mockAccount;
-        }
-        throw error.unknownAccount();
-      });
-
-      await directStripeRoutesInstance.sendSubscriptionDeletedEmail(
-        subscription
-      );
-
-      assert.calledWith(
-        directStripeRoutesInstance.stripeHelper.extractInvoiceDetailsForEmail,
-        subscription.latest_invoice
-      );
-
-      if (accountFound || subscriptionAlreadyCancelled) {
-        assert.notCalled(
-          directStripeRoutesInstance.mailer.sendSubscriptionAccountDeletionEmail
-        );
-      } else {
-        const fakeAccount = {
-          email: mockInvoiceDetails.email,
-          uid: mockInvoiceDetails.uid,
-          emails: [{ email: mockInvoiceDetails.email, isPrimary: true }],
-        };
-        assert.calledWith(
-          directStripeRoutesInstance.mailer
-            .sendSubscriptionAccountDeletionEmail,
-          fakeAccount.emails,
-          fakeAccount,
-          mockInvoiceDetails
-        );
-      }
-    };
-
-    it(
-      'does not send a cancellation email on subscription deletion',
-      commonSendSubscriptionDeletedEmailTest(true)
-    );
-
-    it(
-      'sends an account deletion specific email on subscription deletion when account is gone',
-      commonSendSubscriptionDeletedEmailTest(false)
-    );
-
-    it(
-      'does not send a cancellation email on account deletion when the subscription is already cancelled',
-      commonSendSubscriptionDeletedEmailTest(false, true)
-    );
-  });
-
-  describe('getSubscriptions', () => {
-    const formatter = (subs) =>
-      subs.data.map((s) => ({ subscription_id: s.id }));
-
-    describe('when a customer is found', () => {
-      it('returns a formatted version of the customer subscriptions', async () => {
-        const customer = deepCopy(emptyCustomer);
-        const subscription = deepCopy(subscription2);
-        customer.subscriptions.data = [subscription];
-
-        directStripeRoutesInstance.stripeHelper.customer.resolves(customer);
-        directStripeRoutesInstance.stripeHelper.subscriptionsToResponse.resolves(
-          formatter(customer.subscriptions)
-        );
-
-        const expected = formatter(customer.subscriptions);
-        const actual = await directStripeRoutesInstance.getSubscriptions(
-          VALID_REQUEST
-        );
-
-        assert.deepEqual(expected, actual);
-        assert.calledOnce(
-          directStripeRoutesInstance.stripeHelper.subscriptionsToResponse
-        );
-      });
-    });
-
-    describe('when a customer is not found', () => {
-      it('returns an empty array', async () => {
-        directStripeRoutesInstance.stripeHelper.customer.resolves(null);
-
-        const expected = [];
-        const actual = await directStripeRoutesInstance.getSubscriptions(
-          VALID_REQUEST
-        );
-
-        assert.deepEqual(expected, actual);
-        assert.notCalled(
-          directStripeRoutesInstance.stripeHelper.subscriptionsToResponse
         );
       });
     });
